changed the package.json and moved test out of src
