+++ __tests__/client/ReadWriteBaseClient.test.ts
@@ -1,6 +1,6 @@
-import { ReadWriteBaseClient } from '../../client/ReadWriteBaseClient';
-import RequestMaker from '../../client-helpers/RequestMaker';
-import { RequestArgs } from '../../types';
+import { ReadWriteBaseClient } from '../../src/client/ReadWriteBaseClient';
+import RequestMaker from '../../src/client-helpers/RequestMaker';
+import { RequestArgs } from '../../src/types';
 
 jest.mock('../../client-helpers/RequestMaker');
