+++ package.json
@@ -54,7 +54,8 @@
     "preset": "ts-jest",
     "testEnvironment": "node",
     "roots": [
-      "<rootDir>/src"
+      "<rootDir>/src",
+      "<rootDir>/__tests__"
     ],
     "transform": {
       "^.+\\.tsx?$": "ts-jest"
